Update README.md
# Lido Finance: Liquid Staking For All

![lido vibe cover](/assets/lido-vibe.jpg)

Hey hey, welcome to the Lido GitHub 🔨

The Lido GitHub is a place where Lido contributors come together to grow a wider staking ecosystem 🌍 We are a community of contributors strongly aligned with open-source culture and we welcome any new contributions, feature requests, issues or pull requests 💡

We are excited and honoured to be a part of the web3 developer community - we’ll see you around!

## 🏖️ About Lido

Lido is a liquid-staking protocol built to make staking as secure and decentralized as possible. As a family of open-source, peer-to-system software tools, Lido is currently deployed and functioning on the Ethereum, Solana, and Polygon blockchain networks.

The Lido protocol is governed by the Lido DAO - an independent collection of contributors, staking enthusiasts and governance participants 🏝️

## 🚀 Getting Started

Any and all contributions are greatly appreciated so please do not hesitate to get involved ⚡  
To get started, check out a few of the repos below:

- [lido-dao](https://github.com/lidofinance/lido-dao) - the core Lido protocol,
- [scripts](https://github.com/lidofinance/scripts) - Lido DAO on-chain operations,
- [lido-oracle](https://github.com/lidofinance/lido-oracle) - Lido Oracle daemon,
- [docs](https://github.com/lidofinance/docs) - Lido on Ethereum docs,
- [lido-ethereum-sdk](https://github.com/lidofinance/lido-ethereum-sdk) - Lido Ethereum SDK,
- [ui](https://github.com/lidofinance/ui) - Lido UI library,
- [lido-frontend-template](https://github.com/lidofinance/lido-frontend-template) - a frontend template for Lido web apps.

If it’s of interest, Lido DAO has a [LEGO grants program](https://lido.fi/lego) to reward external contributors for their efforts.  
Check it out and get in touch if you have an idea you want to see built 🧱

## 🔗 Additional Resources

Some of the more useful resources have been compiled below. Browse away 📜

- [Website](https://lido.fi/)
- [Twitter](https://twitter.com/lidofinance)
- [YouTube](https://www.youtube.com/c/LidoFinance)
- [Documentation](https://docs.lido.fi/)
- [Lido DAO Governance and Research forum](https://research.lido.fi/)
- [Bug Bounty](https://lido.fi/bug-bounty)
- [Discord](https://discord.com/invite/lido)

Lastly - if you’re looking to join the team of Lido contributors, check out the [opportunities](https://careers.lido.fi/) page!